requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.2
requests==2.31.0
psycopg2-binary==2.9.7
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2

# Dependências para análise estatística R
# R packages: readr, dplyr, ggplot2, lubridate, forecast

# Dependências do servidor Flask
flask==2.3.3

# Dependências para conexão com banco
oracledb==1.4.2
